docs(event-preview): rewrite stories in new format (#562)
