+++ src/components/EventPreview/stories/basic.story.mdx
@@ -1,0 +1,233 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { EventPreview } from '../EventPreview';
+import { 
+  decorators,
+  onShowDetails,
+  strings,
+  customStyle,
+  exampleTheme,
+} from './helper';
+import { ThemeProvider } from 'styled-components';
+
+<Meta title="Others|EventPreview" decorators={decorators} />
+
+# Event Preview
+
+`EventPreview` component fetches information about an event and presents it as a block element with following items: event type / event subtype,
+event description, event start datetime and end datetime, number of elements in metadata (additional pieces of data), button to explore 
+additional event details. The component requires only `eventId` that is numerical ID of an event. While loading an event the component 
+shows loading spinner.
+
+**NOTE:** The component should have `ClientSDKProvider` as a parent component in react component tree.
+### Basic usage:
+
+Imports:
+
+```jsx
+import 'antd/dist/antd.css';
+
+import React from 'react';
+import { CogniteEvent } from '@cognite/sdk';
+import { EventPreview } from '@cognite/gearbox';
+
+```
+```jsx
+const onShowDetails = (e: CogniteEvent) => {
+  console.log('onShowDetails', e);
+};
+
+```
+<Preview>
+  <Story name="Basic Usage" >
+    <EventPreview eventId={25496029326330} onShowDetails={onShowDetails} />
+  </Story>
+</Preview>
+
+## Available props
+
+<Props of={EventPreview} />
+
+### Types
+
+#### CogniteEvent
+`CogniteEvent` type can be imported from `@cognite/sdk`:
+
+```jsx
+import { CogniteEvent } from '@cognite/sdk';
+```
+
+#### EventPreviewStyles
+This interface defines inline CSS styles for inner elements of `EventPreview` component.
+You can override styles of following blocks:
+
+<img src="event_preview/styling_schema.jpg" alt="EventPreview Styling" width="600px"/>
+<br/><br/>
+
+The type can be imported from `@cognite/gearbox`:
+
+```jsx
+import { EventPreviewStyles } from '@cognite/gearbox';
+```
+
+Definition:
+
+```jsx
+interface EventPreviewStyles {
+  wrapper?: React.CSSProperties;
+  eventType?: React.CSSProperties;
+  title?: React.CSSProperties;
+  button?: React.CSSProperties;
+  times?: React.CSSProperties;
+  metadata?: React.CSSProperties;
+}
+```
+
+See more details in `Custom Styles` example.
+
+## Examples
+
+### Hidden Type
+
+<Preview>
+  <Story name="Hidden Type" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      hideProperties={['type']}
+    />
+  </Story>
+</Preview>
+
+### Hidden Description
+
+<Preview>
+  <Story name="Hidden Description" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      hideProperties={['description']}
+    />
+  </Story>
+</Preview>
+
+### Hidden Start/End Times
+
+<Preview>
+  <Story name="Hidden Start/End Times" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      hideProperties={['startTime', 'endTime']}
+    />  
+  </Story>
+</Preview>
+
+### Hidden Metadata
+
+<Preview>
+  <Story name="Hidden Metadata" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      hideProperties={['metadata']}
+    />  
+  </Story>
+</Preview>
+
+### Hidden Details Button
+
+<Preview>
+  <Story name="Hidden Details Button" >
+    <EventPreview eventId={25496029326330} />  
+  </Story>
+</Preview> 
+
+### With Missing Properties
+
+<Preview>
+  <Story name="With Missing Properties" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+    />  
+  </Story>
+</Preview>
+
+### Without Loading Spinner
+
+<Preview>
+  <Story name="Without Loading Spinner" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      hideLoadingSpinner={true}
+    />  
+  </Story>
+</Preview>  
+
+### With Custom Text
+
+```jsx
+const strings = {
+  start: 'From',
+  end: 'To',
+  details: 'More Details',
+  metadataSummary: 'Contains {{count}} more',
+};
+```
+
+<Preview>
+  <Story name="With Custom Text" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      strings={strings}
+    />  
+  </Story>
+</Preview>
+
+### With Custom Styles
+
+```jsx
+import { EventPreviewStyles } from '@cognite/gearbox';
+
+const customStyle: EventPreviewStyles = {
+  wrapper: { backgroundColor: 'pink' },
+  eventType: { color: 'green' },
+  description: { color: 'yellow' },
+  button: { color: 'black', backgroundColor: 'magenta' },
+  times: { backgroundColor: 'purple' },
+  metadata: { backgroundColor: 'lightblue' },
+};
+```
+
+<Preview>
+  <Story name="With Custom Styles" >
+    <EventPreview
+      eventId={25496029326330}
+      onShowDetails={onShowDetails}
+      styles={customStyle}
+    />  
+  </Story>
+</Preview>
+
+### With Theme
+
+```jsx
+import { ThemeProvider } from 'styled-components';
+
+const exampleTheme = {
+  gearbox: {
+    containerColor: 'AliceBlue',
+    containerBorderColor: 'Aqua',
+    textColorAccent: 'Coral',
+    textColorSecondary: 'ForestGreen',
+  },
+};
+```
+
+<Preview>
+  <Story name="With Theme" >
+    <ThemeProvider theme={exampleTheme}><EventPreview eventId={25496029326330} onShowDetails={onShowDetails}/></ThemeProvider>
+  </Story>
+</Preview>
