--- src/components/EventPreview/stories/hideButton.md
@@ -1,22 +1,0 @@
-## Hidden Details Button
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { EventPreview } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-
-  return (
-    <EventPreview 
-      eventId={4650652196144007}
-    />
-  );
-
-}
-```
