--- src/components/EventPreview/stories/hideDescription.md
@@ -1,26 +1,0 @@
-## Hidden Event Description 
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { CogniteEvent } from '@cognite/sdk';
-import { EventPreview } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  const onShowDetails = (event: CogniteEvent) = {};
-
-  return (
-    <EventPreview 
-      eventId={4650652196144007}
-      onShowDetails={onShowDetails}
-      hideProperties={['description']}
-    />
-  );
-
-}
-```
