--- src/components/EventPreview/stories/withCustomText.md
@@ -1,31 +1,0 @@
-## With Custom Text
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { CogniteEvent } from '@cognite/sdk';
-import { EventPreview } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  const onShowDetails = (event: CogniteEvent) = {};
-
-  return (
-    <EventPreview 
-      eventId={4650652196144007}
-      onShowDetails={onShowDetails}
-      strings={{
-          start: 'From',
-          end: 'To',
-          details: 'More Details',
-          metadataSummary: 'Contains {{count}} more',
-        }}
-    />
-  );
-
-}
-```
